requirements.txt
aiohttp==3.9.5
APScheduler==3.10.4
python-dotenv
requests==2.31.0
py-cord==2.6.1